m20260304

+++ Stackbricks/src/main/java/org/aquamarine5/brainspark/stackbricks/StackbricksVersionData.kt
@@ -23,5 +23,5 @@ interface StackbricksVersionData {
     )
     val forceInstall: Boolean
     val isStable: Boolean
-    val forceInstallLessVersion: Int
+    val forceInstallLessVersion: Long
 }
